Move prev_chal_evals to RecursionChallenge::evals
