+++ kimchi/src/verifier.rs
@@ -20,8 +20,7 @@ use crate::{
 use ark_ff::{Field, One, PrimeField, Zero};
 use ark_poly::{EvaluationDomain, Polynomial};
 use commitment_dlog::commitment::{
-    b_poly, b_poly_coefficients, combined_inner_product, BatchEvaluationProof, CommitmentCurve,
-    Evaluation, PolyComm,
+    combined_inner_product, BatchEvaluationProof, CommitmentCurve, Evaluation, PolyComm,
 };
 use itertools::izip;
 use oracle::{sponge::ScalarChallenge, FqSponge};
@@ -63,51 +62,6 @@ impl<G: CommitmentCurve> ProverProof<G>
 where
     G::BaseField: PrimeField,
 {
-    pub fn prev_chal_evals(
-        &self,
-        index: &VerifierIndex<G>,
-        evaluation_points: &[G::ScalarField],
-        powers_of_eval_points_for_chunks: &[G::ScalarField],
-    ) -> Vec<Vec<Vec<G::ScalarField>>> {
-        self.prev_challenges
-            .iter()
-            .map(|RecursionChallenge { chals, comm: _ }| {
-                // No need to check the correctness of poly explicitly. Its correctness is assured by the
-                // checking of the inner product argument.
-                let b_len = 1 << chals.len();
-                let mut b: Option<Vec<G::ScalarField>> = None;
-
-                (0..2)
-                    .map(|i| {
-                        let full = b_poly(chals, evaluation_points[i]);
-                        if index.max_poly_size == b_len {
-                            return vec![full];
-                        }
-                        let mut betaacc = G::ScalarField::one();
-                        let diff = (index.max_poly_size..b_len)
-                            .map(|j| {
-                                let b_j = match &b {
-                                    None => {
-                                        let t = b_poly_coefficients(chals);
-                                        let res = t[j];
-                                        b = Some(t);
-                                        res
-                                    }
-                                    Some(b) => b[j],
-                                };
-
-                                let ret = betaacc * b_j;
-                                betaacc *= &evaluation_points[i];
-                                ret
-                            })
-                            .fold(G::ScalarField::zero(), |x, y| x + y);
-                        vec![full - (diff * powers_of_eval_points_for_chunks[i]), diff]
-                    })
-                    .collect()
-            })
-            .collect()
-    }
-
     /// This function runs the random oracle argument
     pub fn oracles<
         EFqSponge: Clone + FqSponge<G::BaseField, G, G::ScalarField>,
@@ -309,8 +263,12 @@ where
         let polys: Vec<(PolyComm<G>, _)> = self
             .prev_challenges
             .iter()
-            .zip(self.prev_chal_evals(index, &evaluation_points, &powers_of_eval_points_for_chunks))
-            .map(|(challenge, evals)| {
+            .map(|challenge| {
+                let evals = challenge.evals(
+                    index.max_poly_size,
+                    &evaluation_points,
+                    &powers_of_eval_points_for_chunks,
+                );
                 let RecursionChallenge { chals: _, comm } = challenge;
                 (comm.clone(), evals)
             })
